util.exe_exists and hooks._which were almost identical; merge them
